Allow users to register a username
When registering on the gateway, users can now choose a username. That name
can be used later, e.g. for sending money to another account directly, or
to allow communication with other people through <username>@gateway.jid.

For now, a user can:
  - choose a username when registering
  - change their username when already registered

It is OK to have an empty username. It will mean that you don't want to be
known to the rest of the world.