Payment orders can now go to users
In addition to payments to a Bitcoin address, it's now possible to do
payments to a given username.

On construction, a PaymentOrder accepts either an address or a username.
You can't give both (well, you can, but you'll be yelled an exception at).
In the database, the recipient is represented by a single column.

Payments to usernames result in a 'move' operation inside the bitcoin
wallet: no transaction will be shown outside, and the confirmation is
immediate.