Don't list the gateway to registered users
The only point of listing the gateway itself in service discovery
is to allow users to register, because some clients (e.g. Gajim) only
allow registration on list items, not the browsed element, so typing
the JID of the component directly doesn't allow to register.

So registered users have no need to see that strange self-pointed
element. Hide it.