Don't allow dots in usernames
The node of a local JID is either the username or the real, full JID of
the user. This means there can be a conflict if a user registers with a
JID that corresponds to an existing username.

Since a JID *must* contain a dot, we fix a rule that a username *must not*.
This way, the two namespaces stay separated. Any other character is OK,
even if it contains an @ (or more): the string will always be jid-encoded,
and we never extract a JID from a local JID blindly.