Make sure that JIDs without a dot don't register
Paranoid complement to 9ef242b: don't allow JIDs which contain no dot to
register an account. This should normally not happen, unless e.g. a Jabber
server administrator on the local network tries to register on behalf of
the server itself. Such a behaviour would be dodgy anyway, so we boldly
refuse it, and raise an alert to the component's operator.