Implement jabber:iq:last (XEP-0012) on main JID
Implement the "Last Activity" protocol (XEP-0012), also known as
jabber:iq:last.

At the moment only the main JID (that of the gateway) provides this
information, but it would be easy to update any other JID's last
activity – imply do something like this:
    c.last['xxx'] = datetime.now()
where 'xxx' is the JID whose last presence you want to track, and c is the
component handling the connection.

The gateway's JID is the easiest to track: the spec says that its "last
time" is actually its uptime, so it's never updated.

We won't have to track any "last status message", since all JIDs provided
by the gateway stay alive as long as the gateway is.

With this approach, the only problem is the increasing of the cache, since
it's impossible to tell when e.g. a bitcoin address should disappear.