Override xmpppy's JID by allowing empty domain
xmpppy's JID class forbids the construction of a JID without either a
JID as a string, or a domain. Since in the context of bitcoIM we do know
what our domain is, we can set it to a non-empty value as a class
attribute, and try that value when constructing the JID. Of course, if
the domain was not defined as a class attribute, we can't do anything and
the default constraint applies.

This is a shy step toward the generalization of an "Addressable" class,
from which all addressable objects would inherit (Component, Address,
UserAccount), and that would inherit from JID. I'm not sure whether that's
a good idea though, so this is a careful step in that direction.