Identify JIDs in a separate function
In many occasions, it's important to identify what addressable object
a JID refers to: the component, a user, or a bitcoin address.

Moreover, with the new Addressable class, it's possible that all 3
classes provide the same methods, so it can be useful to generate an
object without caring about its real type, and delegate a certain job
to it.

So the new addressable.generate(jid) function does that. It takes a JID
as an argument and returns a subclass of Addressable.

Note: maybe it's not the better place for doing that job. We'll change
it later when we come up with a better one.