Dispatch IQ handling to Addressable subclasses
The iqHandler() method of the component calls the iqReceived() method of
the right object depending of the "to" attribute.

For now, only implement 2 of the 4 IQ handlers:
  - component handler, that handles registrations and translations from
    a recipient (username, embedded JID, bitcoin address) to a JID.
  - default Addressable handler, that is applied whatever the recipient is.
No handler for UserAccount and Address have been implemented yet.

When a handler terminates without having raised a NodeProcessed event, it
calls the default handler (Addressable.iqReceived).