Allow runtime modification of i18n.fallbackLangs
Sometimes, i18n.fallbackLangs might need to be modified after the i18n
module is loaded. If this variable is used in the signature of the _()
method, it is evaluated only once (when the signature is evaluated),
which prevents further modification of the variable to be reflected in
the behaviour of the method. To avoid this effect, we use None as a
marker, and dynamically use fallbackLangs when this marker is present.